backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
qdrant-client==1.9.0
pydantic==2.5.0
langchain==0.1.15
langchain-community==0.0.31
langchain-google-genai==0.0.11
sqlalchemy==2.0.45
psycopg2-binary==2.9.9